Remove Network
+ AttackObject idea
