+++ Game2/core/src/com/haruham/game/item/Weapon.java
@@ -34,8 +34,9 @@ public class Weapon extends Equipment {
 
     public void release(World world, Character user, Vector3 target) {
             //target = new Vector3(target);
-           // world.addAttack(new AttackObject(user, new MeleeAttack(), new KnockbackCollision(), target, damage));
-            world.addAttack(new AttackObject(user, new Projectile(64), new BasicCollision(), target, damage));
+
+        world.addAttack(new AttackObject(user, new MeleeAttack(), new KnockbackCollision(), target, damage));
+           // world.addAttack(new AttackObject(user, new Projectile(64), new BasicCollision(), target, damage));
 
     }
